[poll] Rename models, squash migrations
