+++ server/venueless/core/models/__init__.py
@@ -12,7 +12,7 @@ from .exhibitor import (
 )
 from .feedback import Feedback
 from .janus import JanusServer
-from .polls import Poll, PollOption, PollVote
+from .poll import Poll, PollOption, PollVote
 from .question import Question, QuestionVote
 from .room import Reaction, Room, RoomView
 from .roulette import RoulettePairing, RouletteRequest
