+++ server/venueless/live/channels.py
@@ -6,5 +6,5 @@ GROUP_ROOM = "room.{id}"
 GROUP_ROOM_QUESTION_READ = "room.{id}.question.read"
 GROUP_ROOM_QUESTION_MODERATE = "room.{id}.question.moderate"
 GROUP_ROOM_POLL_READ = "room.{id}.poll.read"
-GROUP_ROOM_POLL_MODERATE = "room.{id}.poll.manage"
+GROUP_ROOM_POLL_MANAGE = "room.{id}.poll.manage"
 GROUP_ROULETTE_CALL = "roulette.{id}"
